added new apps for notifications and products, as well as redid some forms with templates
