+++ account/templates/login.html
@@ -9,18 +9,13 @@ Login
 {% endblock %}
 
 {% block content %}
+
 <h1>Login</h1>
 <hr/>
-{% if messages %}
-<ul class="messages">
-    {% for message in messages %}
-    <li>{{ message }}</li>
-    {% endfor %}
-</ul>
-{% endif %}
-
+{% include "message_form_block.html" with messages=messages %}
 {% include "account_forms.html" with form=form submittext="Login" includereq=False %}
 <br/>
-<a href="/signup/">Don't have an account yet?</a><br/>
-<a href="/password_reset/">Forgot your password?</a>
+<a href="{% url 'signup' %}">Don't have an account yet?</a><br/>
+<a href="{% url 'password_reset' %}">Forgot your password?</a>
+
 {% endblock %}
