+++ account/templates/password_management/password_change.html
@@ -7,15 +7,9 @@ Change Password
 {% block content %}
 <h1>Change Password</h1>
 <hr/>
-{% if messages %}
-<ul class="messages">
-    {% for message in messages %}
-    <li>{{ message }}</li>
-    {% endfor %}
-</ul>
-{% endif %}
 
+{% include "message_form_block.html" with messages=messages %}
 {% include "account_forms.html" with form=form submittext="Submit" includereq=False %}
 <br/>
-<a href="/management/"><button type="submit">Back</button></a>
+<a href="{% url 'management' %}"><button type="submit">Back</button></a>
 {% endblock %}
