+++ account/templates/password_management/password_reset.html
@@ -4,17 +4,10 @@
 Forgot Your Password?
 {% endblock %}
 
-{% block nav %}
-
-{% endblock %}
-
 {% block content %}
+
   <h1>Forgot your password?</h1>
   <p>Enter your email address below, and we'll email instructions for setting a new one.</p>
+  {% include "account_forms.html" with form=form submittext="Send me instructions!" includereq=False %}
 
-  <form method="POST">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <input type="submit" value="Send me instructions!">
-  </form>
 {% endblock %}
