+++ account/templates/password_management/password_reset_complete.html
@@ -1,11 +1,7 @@
 {% extends 'base.html' %}
 
-{% block nav %}
-
-{% endblock %}
-
 {% block content %}
   <p>
-    Your password has been set. You may go ahead and <a href="{% url 'login' %}">sign in</a> now.
+    Your password has been set.
   </p>
 {% endblock %}
