+++ account/templates/password_management/password_reset_confirm.html
@@ -5,17 +5,16 @@
 {% endblock %}
 
 {% block content %}
-  {% if validlink %}
-    <h3>Change password</h3>
-    <form method="post">
-      {% csrf_token %}
-      {{ form.as_p }}
-      <button type="submit">Change password</button>
-    </form>
-  {% else %}
-    <p>
-      The password reset link was invalid, possibly because it has already been used.
-      Please request a new password reset.
-    </p>
-  {% endif %}
+
+{% if validlink %}
+<h3>Change password</h3>
+{% include "account_forms.html" with form=form submittext="Change Password" includereq=False %}
+
+{% else %}
+<p>
+  The password reset link was invalid, possibly because it has already been used.
+  Please request a new password reset.
+</p>
+{% endif %}
+
 {% endblock %}
