+++ account/templates/password_management/password_reset_done.html
@@ -4,11 +4,8 @@
 Forgot Your Password?
 {% endblock %}
 
-{% block nav %}
-
-{% endblock %}
-
 {% block content %}
+
   <p>
     We've emailed you instructions for setting your password, if an account exists with the email you entered.
     You should receive them shortly.
@@ -17,4 +14,5 @@ Forgot Your Password?
     If you don't receive an email, please make sure you've entered the address you registered with,
     and check your spam folder.
   </p>
+
 {% endblock %}
