+++ account/templates/signup.html
@@ -12,5 +12,8 @@ Login
 <h1>Account Registration</h1>
 <hr/>
 <p><b>*</b> fields are required!</p>
+{% include "message_form_block.html" with messages=messages %}
 {% include "account_forms.html" with form=form submittext="Sign Up" includereq=True %}
+<br/>
+<a href="{% url 'login' %}">Already have an account?</a><br/>
 {% endblock %}
