--- core/forms.py
@@ -1,5 +1,0 @@
-from django import forms
-
-
-class CreateDashboardBlock(forms.ModelForm):
-    pass
